requirements.txt
googlesearch-python
pandas
